web/app/requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
bs4==0.0.1
click==6.7
dill==0.2.6
enum34==1.1.6
Flask==0.12
future==0.16.0
futures==3.0.5
gapic-google-cloud-datastore-v1==0.14.1
gapic-google-cloud-logging-v2==0.90.1
gapic-google-cloud-pubsub-v1==0.14.1
google-auth==0.5.0
google-auth-httplib2==0.0.2
google-cloud==0.22.0
google-cloud-bigquery==0.22.1
google-cloud-bigtable==0.22.0
google-cloud-core==0.22.1
google-cloud-datastore==0.22.1
google-cloud-dns==0.22.0
google-cloud-error-reporting==0.22.0
google-cloud-happybase==0.22.0
google-cloud-language==0.22.2
google-cloud-logging==0.22.0
google-cloud-monitoring==0.22.0
google-cloud-pubsub==0.22.0
google-cloud-resource-manager==0.22.0
google-cloud-runtimeconfig==0.22.0
google-cloud-storage==0.22.0
google-cloud-translate==0.22.0
google-cloud-vision==0.22.0
google-gax==0.15.4
googleapis-common-protos==1.5.0
grpc-google-cloud-datastore-v1==0.14.0
grpc-google-cloud-logging-v2==0.90.0
grpc-google-cloud-pubsub-v1==0.14.0
grpc-google-iam-v1==0.11.1
grpcio==1.1.0
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.9.5
mailjet-rest==1.2.2
MarkupSafe==0.23
oauth2client==3.0.0
packaging==16.8
ply==3.8
protobuf==3.2.0
pyasn1==0.2.1
pyasn1-modules==0.0.8
pyparsing==2.1.10
requests==2.11.1
retrying==1.3.3
rsa==3.4.2
six==1.10.0
Werkzeug==0.11.15
